Strict type for data & move dispatchState to utils
